+++ stencil-workspace/src/components/modus-message/modus-message.tsx
@@ -11,7 +11,7 @@ import { IconMap } from '../icons/IconMap';
 })
 export class ModusMessage {
   /** (optional) The message's aria-label. */
-  @Prop() ariaLabel: string;
+  @Prop() ariaLabel: string | null;
 
   /** (optional) The message's icon. */
   @Prop() icon?: string;